Feat/backend: refactor response models to use UUIDs and OffsetDateTime for improved data handling

+++ backend/src/main/java/org/codeup/parknexus/service/impl/UserServiceImpl.java
@@ -50,7 +50,7 @@ public class UserServiceImpl implements IUserService {
                 .activeReservations(activeReservations)
                 .activeSessions(activeSessions)
                 .totalReservations(totalReservations)
-                .outstandingFees(outstandingFees.doubleValue())
+                .outstandingFees(outstandingFees)
                 .build();
     }
 }
