+++ backend/src/main/java/org/codeup/parknexus/web/dto/admin/AdminDashboardResponse.java
@@ -5,11 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data @NoArgsConstructor @AllArgsConstructor @Builder
 public class AdminDashboardResponse {
     private Long totalSpots;
     private Long occupiedSpots;
-    private Double totalRevenue;
+    private BigDecimal totalRevenue;
     private Long totalUsers;
     private Long adminsCount;
     private Long usersCount;
